pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "RAG知识库问答系统 - 基于LangChain、FastAPI、Chroma和MongoDB构建"
authors = [
    {name = "hbchen", email = "hbchen7@126.com"},
]

# 核心依赖 - 运行时必需
dependencies = [
    # Web 框架
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "starlette",
    "pydantic",
    # 配置和环境
    "python-dotenv>=1.1.0",
    # 数据库相关
    "motor>=3.7.0",
    "beanie>=1.29.0",
    "pymongo",
    # LangChain 核心
    "langchain>=0.3.21",
    "langchain-core",
    "langchain-community>=0.3.20",
    "langchain-chroma>=0.2.2",
    "langchain-openai>=0.3.12",
    "langchain-ollama>=0.3.0",
    "langchain-mongodb>=0.6.0",
    "langchain-text-splitters",
    # 文档处理 - 基础版本
    "pypdf>=5.4.0",
    "python-docx>=1.1.2",
    "markdown>=3.8",
    "unstructured[docx]>=0.17.2", # 高级文档处理
    # 安全相关
    "python-jose[cryptography]>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    # HTTP 客户端
    "httpx",
    "aiohttp>=3.11.18",
    # 搜索相关
    "rank-bm25>=0.2.2",
]

requires-python = ">=3.10"
license = {text = "MIT"}

# 可选依赖组 - 按功能分组
[dependency-groups]
# 开发依赖
dev = [
    "ruff>=0.11.4",   
]
# 测试
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0", # 支持异步测试
    "pytest-html>=4.1.1", # HTML测试报告
    "pytest-cov>=6.0.0", # 测试覆盖率
    "locust>=2.34.1", #压力测试
]

# RAG评估
ragas = [
    "datasets>=4.1.1",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "pyyaml>=6.0.2",
    "ragas>=0.3.5",
    "seaborn>=0.13.2",
]

# Jupyter Notebook 支持 - 临时开发使用
jupyter = [
    "jupyter>=1.1.1",           # Jupyter核心
    "jupyterlab>=4.3.4",        # JupyterLab界面
    "notebook>=7.3.1",          # Jupyter Notebook
    "ipykernel>=6.29.5",        # Python内核
    "ipywidgets>=8.1.5",        # 交互式组件
    "matplotlib>=3.10.0",       # 绘图库
    "seaborn>=0.13.2",          # 统计绘图
    "plotly>=5.25.0",           # 交互式图表
]

# Agent 功能 - 可选安装
agent = [
    "langchain-mcp-adapters>=0.1.0",
    "fastapi-mcp>=0.3.3",
    "langchain-tavily>=0.1.6",
]

# AI/ML 功能 - 可选安装
ai-full = [
    "sentence-transformers>=4.1.0",    # 大模型，约 2GB - 仅在需要时安装
    "bert-score>=0.3.13",              # BERT 相关，约 1GB - 仅在需要时安装
    "nltk>=3.9.1",
]

# 文档分块-语义分割功能 - 可选安装
semantic = [
    "langchain-experimental>=0.3.4",   # SemanticChunker 需要
]


#  弃用langgraph
discard-agent = [
    "langgraph>=0.3.34",
    "langgraph-checkpoint-mongodb>=0.2.0",

]

# tortoise数据库迁移 - 开发时使用
migration = [
    "tortoise-orm>=0.24.2",
    "asyncpg>=0.30.0",
    "aerich>=0.8.2",
    "tomlkit>=0.13.2",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.aerich]
tortoise_orm = "settings.PostgreSQLConfig"
location = "./migrations"
src_folder = "./."

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.ruff]
select = ["F", "I"]
line-length = 88
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.deptry.per_rule_ignores]
DEP001 = ["src"]
DEP002 = [
    "python-multipart",
    "tortoise-orm",
    "asyncpg",
    "aerich",
    "tomlkit",
    "pypdf",
    "markdown",
    "python-docx",
    "aiohttp",
    "rank-bm25",
]
